refactor(cst): node-side utilities move out of file_analysis
node_to_span / extract_call_name / fq_tail_span / find_ancestor now
live in cst.rs (re-exported from file_analysis while the lazy
tree-resolution paths — phase 5 retires them — still call them
in-module); count_call_args is cst::call_args().len(). What remains of
tree-sitter in file_analysis.rs is Point/Tree signature types plus the
three lazy walkers (resolve_expression_type's degradation walk,
resolve_hash_owner_from_tree, call_arg_key_at), which are called from
find_definition/collect_refs_for_target's lazy paths and can only move
when eager Ref.target lands.
